pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "f19", email = "80933556+ff19@users.noreply.github.com"},
]
dependencies = [
    "graia-saya==0.0.15",
    "graia-ariadne==0.6.15",
    "requests==2.27.1",
    "lxml>=4.9.0",
    "pymongo>=4.1.1",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
